Changed BitArray from IntArray to LongArray

+++ kbloom/src/main/java/com/nguyenmoclam/kbloom/core/BloomFilterData.kt
@@ -10,5 +10,5 @@ data class BloomFilterData(
     val bitSetSize: Int,
     val numHashFunctions: Int,
     val seed: Int,
-    val bitArray: List<Int>
+    val bitArray: List<Long>
 )
